Add python impl

+++ OrderMatching/OrderBook.h
@@ -49,7 +49,7 @@ private:
 inline void OrderBook::countFilledCount() {
     uint64_t count = 0;
     uint64_t totalVolume = 0;
-    for (int i = 0; i < orderCount; i++) {
+    for (uint64_t i = 0; i < orderCount; i++) {
         if (orders[i] != nullptr && orders[i]->getFilledVolume() == orders[i]->getOriginalVolume()) {
             totalVolume+= orders[i]->getFilledVolume();
             count++;
